Task #6076 for testing
Modified AI so that it now looks like it’s playing along.  Button
appear to be clicked and the clicking now inverts.  The buttons when
being used by the AI still can be clicked by the user to cause them to
invert, but no functionality executes meaning it will not affect game
play.  Also some slight tweaks to a few CSV files.